// src/app/shared/model/user.ts
export class User{
    username?: string;
    nombreApellidos?: string;
    phone?: string;
    edad?: string;
    uri?: string;
    localidad?: string;
    playlists?: string[];

}

export class PlayList{
push(data: PlayList) {
  throw new Error('Method not implemented.');
}
nombre?: string;
// totalTracks?: string;
description?: string;
tracks?: string[];
}